Errors: shell, exceptions

+++ core/source/userdata.cpp
@@ -1,5 +1,6 @@
 #include "userdata.h"
 #include "constants.h"
+#include "validators.h"
 #include <filesystem>
 
 
@@ -9,8 +10,10 @@ namespace mb::core
 
 	std::vector<std::string> getPlayLists(const int64_t &user_id)
 	{
+		constexpr auto sl = "/";
+
 		std::vector<std::string> res;
-		std::string target_dir = mb::pth::USER_DATA_DIR + std::to_string(user_id) + '/';
+		std::string target_dir = mb::pth::USER_DATA_DIR + std::to_string(user_id) + sl;
 
 		if (!fs::exists(target_dir)) {
 			fs::create_directory(target_dir);
@@ -54,9 +57,14 @@ namespace mb::core
 
 	void renamePlayList(const int64_t &user_id, const std::string &old_name, const std::string &new_name) 
 	{
-		constexpr auto sl = "/";
+		constexpr auto sl  = "/";
+
 		std::string pthold = pth::USER_DATA_DIR + std::to_string(user_id) + sl + old_name;
 		std::string pthnew = pth::USER_DATA_DIR + std::to_string(user_id) + sl + new_name;
+
+		if (!fs::exists(pthold)) {
+			throw err::NOT_EXISTED_PLAYLIST;
+		}
 		
 		fs::rename(pthold, pthnew);
 	}
@@ -64,6 +72,7 @@ namespace mb::core
 	void removePlayList(const int64_t &user_id, const std::string &name) 
 	{
 		constexpr auto sl = "/";
+
 		const std::string path = pth::USER_DATA_DIR + std::to_string(user_id) + sl + name;
 
 		if (!fs::exists(path)) {
@@ -71,5 +80,20 @@ namespace mb::core
 		}
 
 		fs::remove_all(pth::USER_DATA_DIR + std::to_string(user_id) + sl + name);
+	}
+
+	bool protectedShell(const int64_t &id, TgBot::Bot &bot, const std::function<void()> &func)
+	{
+		static const std::string text = core::parseHTML(pth::MESSAGE_DIR + pth::HTML_OLD_DATA);          // parse only one
+		
+		try {
+			func();
+		}
+		catch (const BotError &err) {
+			bot.getApi().sendMessage(id, text, false, 0, nullptr, mrk::HTML);
+			return false;
+		}
+
+		return true;
 	}
 }
